screen-recorder/requirements.txt
numpy==1.17.4
pyscreenshot==0.5.1
opencv-python==4.1.2.30
Pillow==6.2.1